requirements.txt
fastapi>=0.118.0
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
motor==3.6.1          # async MongoDB driver (latest stable)
pydantic>=2.12.0
langchain>=0.3.27     # works with both OpenAI/Groq wrappers
langchain-openai>=0.3.35  # required for langchain_openai imports
openai>=1.104.2       # compatible version for langchain-openai
tavily-python         # Enhanced search for travel planning
